Separates classes into files, for better reading
