+++ Sources/AnimationSequence/AnimationConfiguration.swift
@@ -1,25 +1,7 @@
 import SwiftUI
 
-// enum that maps the SwiftUI.Animation options
-public enum AnimationEasing {
-    case `default`
-    case easeIn
-    case easeOut
-    case easeInOut
-    case linear
-    case custom(animation: Animation)
-}
-
 // Used to provide animation blocks, and any type of callback
 public typealias AnimationBlock = () -> Void
-
-// MARK: - Default values
-
-public struct AnimationDefaults {
-    public static let defaultDelay: Double = 0
-    public static let defaultDuration: Double = 0.1
-    public static let defaultEasing: AnimationEasing = .default
-}
 
 public class AnimationConfiguration: CustomStringConvertible {
     // MARK: - Properties
