+++ Sources/AnimationSequence/AnimationDefaults.swift
@@ -1,0 +1,12 @@
+//
+//  AnimationDefaults.swift
+//  AnimationSequence
+//
+//  Created by Cristhian Leon on 1/6/25.
+//
+
+public enum AnimationDefaults {
+    public static let defaultDelay: Double = 0
+    public static let defaultDuration: Double = 0.1
+    public static let defaultEasing: AnimationEasing = .default
+}
